ml/speech-to-text/requirements.txt
azure-cognitiveservices-speech==1.40.0
azure-storage-blob==12.19.0
ffmpeg-python==0.2.0
fastapi==0.104.1
moviepy==1.0.3
pydantic==2.5.0
python-dotenv==1.0.0
uvicorn==0.24.0
Hypercorn==0.16.0
